RPM: Added link with version-less jar-filename.
This solves the problem created by removing the finalName-element in a previous commit. The
finalName used to be set to only the artifactId (so without the version number). This way the
shell script for starting the command line interface and the service can stay the same over versions.
However, this led to other problems. By linking to the jar-file that contains the version in its
name we can have the best of both worlds